:wrench: TeamParticipant class moved to its own file.
This commit and the last one close issue #108
