+++ eldonationtracker/tests/test_team.py
@@ -1,5 +1,5 @@
 """Unit tests for team.py."""
-
+import eldonationtracker.team_participant
 from eldonationtracker import team as team
 
 
@@ -31,5 +31,5 @@ def test_str_no_json_data_no_team_id():
 def test_TeamParticipant_str():
     team_donor1_json = {"displayName": "donor1", "sumDonations": "45", "donorID": 1000111,
                         'avatarImageURL': "http://someplace.com/image.jpg", "numDonations": 2}
-    team_donor1 = team.TeamParticipant(team_donor1_json)
+    team_donor1 = eldonationtracker.team_participant.TeamParticipant(team_donor1_json)
     assert str(team_donor1) == "A Team Participant named donor1 who has donated $45.00 to the team over 2 donations."
